feat(apps/web): implement BulkTrackCard and CoverSelectionBanner components for enhanced bulk track upload
- Introduced BulkTrackCard component for displaying and managing individual tracks in the bulk upload form.
- Added CoverSelectionBanner component to facilitate cover art selection from uploaded tracks.
- Created types for BulkTrackUploadForm and related components to improve type safety and clarity.
- Refactored BulkTrackUploadForm to utilize new components and streamline track management logic.
